Moved UserRoleBaseModel to its own file. Added files for base schemas
