+++ libs/core/cornflow_core/models/__init__.py
@@ -6,4 +6,5 @@ from .meta_models import TraceAttributesModel
 from .action import ActionBaseModel
 from .role import RoleBaseModel
 from .user import UserBaseModel
+from .user_role import UserRoleBaseModel
 from .view import ViewBaseModel
